mcp_registry/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.10
pymysql==1.0.3
python-multipart==0.0.6
oauthlib==3.2.2
passlib==1.7.4
python-jose==3.3.0
prometheus-client==0.17.1
prometheus-fastapi-instrumentator==6.0.0
cryptography==41.0.2